requirements.txt
asgiref==3.7.2
base58==2.1.1
Django==4.2.5
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.1
psycopg2==2.9.7
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
